+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.2] - 2019-06-04
+### Fixed
+- Fix `configureTranslation` so that it actually does what it says it does.
+
+## [0.1.1] - 2019-06-04
+### Added
+- Make `DateRangePicker` always fluid when it is inside a Semantic UI Form.
+
 ## [0.1.0] - 2019-05-23
 ### Added
 - First version of this project.